test(modal): speed up sheet tests
